+++ client/index.html
@@ -5,7 +5,7 @@
  <meta charset="UTF-8">
  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
  <meta name="HandheldFriendly" content="true" />
- <link rel="icon" type="image/x-icon" href="./favicon.ico">
+ <link rel="icon" type="image/x-icon" href="./assets/favicon.ico">
  <title>👨🏻‍💻 SimpleChat - mensajes directos | Sergio Ortega dev</title>
