+++ server/src/controllers/chatController.js
@@ -1,0 +1,3 @@
+import chatModels from "../models/chatModels.js";
+
+export default chatModels;
